Enhance supported selectors and increase test coverage
