+++ intern.json
@@ -21,7 +21,7 @@
 				"packages": [
 					{ "name": "src", "location": "_build/src" },
 					{ "name": "tests", "location": "_build/tests" },
-					{ "name": "css-select-umd", "location": "node_modules/css-select-umd", "main": "css-select" },
+					{ "name": "css-select-umd", "location": "node_modules/css-select-umd", "main": "dist/index.js" },
 					{ "name": "diff", "location": "node_modules/diff", "main": "dist/diff.js" }
 				]
 			}
