+++ package.json
@@ -36,10 +36,8 @@
     "@dojo/widget-core": "~0.6.1"
   },
   "dependencies": {
-    "css-select-umd": "file:~/development/dojo2/test-extras/deps/css-select-umd.tgz",
-    "diff": "^3.4.0",
-    "jsdom": "^10.0.0",
-    "pepjs": "^0.4.2",
+    "css-select-umd": "1.3.0-rc0",
+    "diff": "3.4.0",
     "tslib": "~1.8.1"
   },
   "devDependencies": {
@@ -58,7 +56,6 @@
     "prettier": "1.9.2",
     "sinon": "^2.2.0",
     "tslint": "5.8.0",
-    "tslint-plugin-prettier": "1.3.0",
     "typescript": "~2.6.1"
   },
   "lint-staged": {
